Update es-ES.json
The original JSON file contained a bunch of orthographic errors, a few commands untranslated, and was pretty informal and not polite at all. The translation may contain an error, even though I tried to fix it, which is the "Ge", "Gue" and "Güe". I'm politely requesting the change, as I was an hour translating Spanish into... Spanish. Thank you for your comprehension, a Newbie, Daniel Schieffermint. (I don't speak German, but I'm native Spanish.)
Postdata (After-Writing-The-Card Message): Although he tried, it was so unpolite I wrote a bunch of "Thank you" and "Please" after a few sentences. I also have to say, misleading "Hayas" into "Allas" is wild.